+++ NEWS.md
@@ -3,6 +3,7 @@
 - Improve rendering of templates.
 No longer force render data as JSON if using an HTML template.
 - Allow nested partials, their path must be relative.
+- Added `jobj` function to serialise objects to JSON in `render`.
 
 # ambiorix 2.0.0